Add files via upload
This code is extracted from he original version of the code for phase 1. Different logical parts were separated in different files. The merge_datasets file is the first piece of code to load all the CSVs from the server and merge them together while also removing non-CS jobs. Jia will push another file only to find the best keywords out of the job descriptions. That file produces the output that is stored on the server and loaded in the model_training file